requirements.txt
astroid==1.6.2
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-summernote==0.8.8.2
easy-thumbnails==2.4.2
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.1.0
mccabe==0.6.1
nltk==3.2.5
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.3
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
requests==2.18.4
selenium==3.8.0
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
wrapt==1.10.11
